Do not send events to windows after they are released. (And update tests to match.)
